zig-ladder: lift the back-end harness out of the fibx generator
A second rung wants the same harness with a different subject, so the body
moves to emit_harness.py and each rung keeps only the program it compiles.
The walkers take a name prefix so two harnesses cannot collide, and
bundle_fibx.ps1 gains a $PlugName parameter so a caller can label its own
subject.

Provably a no-op for fibx: regenerating FibxHarness.codex after the
extraction produces a byte-identical file.
